+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## [Unreleased]
 * Lib:
     * XML generation:
+        * Allow in device section, devices of type `shmem` and `hostdev`
+
+## [Unreleased]
+* Lib:
+    * XML generation:
         * Allow in feature section, customization of the `kvm` options, in domains
 
 ## [0.4.2]